Added pundit for auth policies. Added policy for age model. Changed application controller to use pundit. Wired up ages_controller to check pundit for permissions

+++ app/controllers/admin/ages_controller.rb
@@ -1,16 +1,20 @@
 class Admin::AgesController < ApplicationController
   before_action :authenticate_user!
-  before_action :require_admin, except: [:index, :show]
 
   def index
     @ages = Age.all.paginate(
       :page => params[:page],
       :per_page => 10
     )
+    authorize @ages
   end
 
   def create
+    # Create instance of age from params
     @age = Age.new(age_params)
+    # Check with pundit if the user has permission
+    authorize @age
+    # Save the new age
     if @age.save
       flash[:notice] = 'Age successfully created'
       redirect_to admin_ages_path
@@ -20,15 +24,25 @@ class Admin::AgesController < ApplicationController
   end
 
   def new
+    # Create new instance of age
     @age = Age.new
+    # Check with pundit if able to create new
+    authorize @age
   end
 
   def edit
+    # Load the age
     @age = Age.find(params[:id])
+    # Check with pundit if can edit
+    authorize @age
   end
 
   def update
+    # Get the age to update
     @age = Age.find(params[:id])
+    # Check with pundit if the user is admin so has permission
+    authorize @age
+    # Survived, so update
     if @age.update(age_params)
       flash[:notice] = 'Age successfully updated'
       redirect_to admin_ages_path
@@ -38,7 +52,11 @@ class Admin::AgesController < ApplicationController
   end
 
   def destroy
+    # Get the age to destroy
     @age = Age.find(params[:id])
+    # Check with pundit if can be destroyed
+    authorize @age
+    # Survived so destroy
     if @age && @age.destroy
       flash[:notice] = 'Agent successfully removed'
     else
